+++ lib/src/internal/new/cells_layout_child.dart
@@ -1,6 +1,5 @@
 import 'package:davi/src/internal/new/cells_layout.dart';
 import 'package:davi/src/internal/new/cells_layout_parent_data.dart';
-import 'package:davi/src/internal/new/row_region.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:meta/meta.dart';
@@ -9,14 +8,15 @@ import 'package:meta/meta.dart';
 class CellsLayoutChild<DATA>
     extends ParentDataWidget<CellsLayoutParentData> {
 
+  static const _CellKey _trailingKey =  _CellKey(childIndex: -100);
+
   factory CellsLayoutChild.cell({required int childIndex, required int rowIndex,
     required int columnIndex, required Widget child}){
    return CellsLayoutChild._(key:_CellKey(childIndex: childIndex), rowIndex: rowIndex, columnIndex: columnIndex, child: child);
   }
 
-  factory CellsLayoutChild.row({required int childIndex, required int rowIndex,
-    required RowRegion<DATA> child}){
-    return CellsLayoutChild._(key:_RowKey(childIndex: childIndex), rowIndex: rowIndex, columnIndex: null, child: child);
+  factory CellsLayoutChild.trailing({required Widget child}){
+    return CellsLayoutChild._(key:_trailingKey, rowIndex: -1, columnIndex: -1, child: child);
   }
 
    const CellsLayoutChild._({
@@ -53,23 +53,6 @@ class CellsLayoutChild<DATA>
     properties.add(DiagnosticsProperty<Object>('rowIndex', rowIndex));
     properties.add(DiagnosticsProperty<Object>('columnIndex', columnIndex));
   }
-}
-
-class _RowKey extends LocalKey {
-
-  const _RowKey({required this.childIndex});
-
-  final int childIndex;
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-          other is _RowKey &&
-              runtimeType == other.runtimeType &&
-              childIndex == other.childIndex;
-
-  @override
-  int get hashCode => childIndex.hashCode;
 }
 
 class _CellKey extends LocalKey {