+++ lib/src/internal/new/cells_layout_parent_data.dart
@@ -7,6 +7,6 @@ class CellsLayoutParentData extends ContainerBoxParentData<RenderBox> {
   int? rowIndex;
   int? columnIndex;
 
-  bool get isCell => columnIndex!=null;
-  bool get isRow => !isCell;
+  bool get isCell => rowIndex!=null && rowIndex!>=0 && columnIndex!=null && columnIndex!>=0;
+  bool get isTrailing => !isCell;
 }